[Release][app] Update build_docker.sh

+++ src/scripts/build_docker.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
 set -e
 
+if [[ ${SRV} == "" || ${DOMAIN} == "" ]]; then
+  exit 1
+fi
+
 cd /home/ci_user/src
 sudo docker builder prune -f && \
 sudo cp config/docker/${DOMAIN}.yml ./docker-compose.yml
@@ -9,27 +13,22 @@ sudo ln -sf /home/ci_user/.certbot ./.certbot
 
 if [[ ${SRV} == *"nginx"* ]]; then
   sudo DOMAIN=${DOMAIN} docker compose build nginx --no-cache
-  sudo DOMAIN=${DOMAIN} docker compose down nginx
-  sudo DOMAIN=${DOMAIN} docker compose up -d nginx
 fi
 
 if [[ ${SRV} == *"app"* ]]; then
   sudo DOMAIN=${DOMAIN} docker compose build recognizer retriever --no-cache
-  sudo DOMAIN=${DOMAIN} docker compose down recognizer retriever
-  sudo DOMAIN=${DOMAIN} docker compose up -d recognizer retriever
 fi
 
 if [[ ${SRV} == *"mongo"* ]]; then
   sudo DOMAIN=${DOMAIN} docker compose build mongo mongo-express --no-cache
-  sudo DOMAIN=${DOMAIN} docker compose down mongo mongo-expres
-  sudo DOMAIN=${DOMAIN} docker compose up -d mongo mongo-express
 fi
 
 if [[ ${SRV} == ",," ]]; then
   sudo docker build --no-cache -t app_base -f ./Dockerfile.base .
   sudo DOMAIN=${DOMAIN} docker compose build --no-cache
-  sudo DOMAIN=${DOMAIN} docker compose down
-  sudo DOMAIN=${DOMAIN} docker compose up -d
 fi
+
+sudo DOMAIN=${DOMAIN} docker compose down
+sudo DOMAIN=${DOMAIN} docker compose up -d
 
 sudo docker system prune -a -f --filter label="can_delete=true" --volumes
